bitcoin_practice_nick/requirements.txt
requests
textblob
tweepy
beautifulsoup4
selenium
praw
Pillow
google_images_download